Moved lib and schemas
