+++ bin/swagger-tools
@@ -29,7 +29,7 @@
 var _ = require('lodash');
 var async = require('async');
 var fs = require('fs');
-var helpers = require('../lib/helpers');
+var helpers = require('../src/lib/helpers');
 var paddingAmount = 18;
 var path = require('path');
 var pkg = require('../package.json');
