+++ middleware/helpers.js
@@ -27,7 +27,7 @@
 /* This module contains code that is reused in more than one of the Swagger middlewares */
 
 var _ = require('lodash');
-var helpers = require('../lib/helpers');
+var helpers = require('../src/lib/helpers');
 var validators = require('../src/lib/validators');
 var parseurl = require('parseurl');
 var qs = require('qs');
