+++ middleware/swagger-metadata.js
@@ -27,7 +27,7 @@
 var _ = require('lodash');
 var async = require('async');
 var bp = require('body-parser');
-var cHelpers = require('../lib/helpers');
+var cHelpers = require('../src/lib/helpers');
 var debug = require('debug')('swagger-tools:middleware:metadata');
 var mHelpers = require('./helpers');
 var multer = require('multer');
