+++ middleware/swagger-validator.js
@@ -26,7 +26,7 @@
 
 var _ = require('lodash');
 var async = require('async');
-var cHelpers = require('../lib/helpers');
+var cHelpers = require('../src/lib/helpers');
 var debug = require('debug')('swagger-tools:middleware:validator');
 var mHelpers = require('./helpers');
 var validators = require('../src/lib/validators');
