+++ src/schemas/1.2/apiDeclaration.json
