+++ src/schemas/1.2/authorizationObject.json
