+++ src/schemas/1.2/dataTypeBase.json
