+++ src/schemas/1.2/infoObject.json
