+++ src/schemas/1.2/resourceListing.json
