+++ src/schemas/1.2/resourceObject.json
